analytics/requirements.txt
psycopg2-binary
pandas
matplotlib
seaborn
sqlalchemy
python-dotenv
fastapi
uvicorn[standard]
